Require micrometer-tracing-bridge-otel to auto-configure OTel
Closes gh-32503
